Make `propagate`, `scc` generic over difference type
The `propagate` and `scc` module currently works for collections with
`R=isize`.  This patch makes all methods of `propagate` generic over
`R`, given the following trait bounds:

```
    R: ExchangeData + Abelian
    R: Mul<R, Output=R>,
    R: From<i8>,
```

The last bound is only needed to convert constants `0` and `1` to `R`
and could be replaced with `R: num::Zero + num::One`; however this would
introduce a new dependency on the `num` crate to differential, which
would only be used here.