Make offset generic in ordered batch implementation
One thing differential does _not_ need more of?  Trait bounds!

This commit generalizes trace and batch implementations based on
sorted ranges (see `ord.rs` and `ordered.rs`) by parameterizing them by
the offset type.  The natural default for this type is `usize`, as it is
used to index into the vector of keys; however this is wasteful in
applications where batch size never exceeds 2^32, where `u32` is
sufficient.

An earlier commit added a separate implementation of `OrdValSpine` and
related types with `u32` offsets, but this involved lots of duplicated
code.  This commit walks back that change and uses type argument
instead.