Require arrangement sharing to communicate frontier changes.
This PR requires that an arrangement's `enter_at` method takes a method to invert timestamps, so that frontier statements about the wrapper trace can be communicated back to the wrapped trace.
This method requires some care, but for the moment you probably shouldn't be using these methods without knowing about what is going on here; we could hope for a future interface that is clearer and less error prone.