Remove the `Default` requirement from keys.
This PR removes the requirement that keys in `lookup_map` implement the `Default` trait.
The implementation was required to get access to instances to re-use in key extraction, but we can instead ask the caller to supply the instances themselves; when a type implements `Default` this is especially easy.
The constraint is only changed on `lookup_map` and not the `count`, `propose`, or `validate` methods.