Balance the effort in join.
The join operator yields after a fixed number of outputs are produced. It does so starting with the queue of batches from its first input, which has the potential to starve the processing of the second input. This starvation results in non-processing and non-compaction of the second trace, which increases the per-batch time to process the first batches. The whole thing has the potential to spiral out of control.

There is still an issue if we fall behind in data processing, as retiring batches is what enables physical compaction. It may be better policy to yield less if the todo lists are becoming long, or to find other ways to structure the backing LSM so that this is less of an issue.

cc @umanwizard