Capture and replay implementations (#279)
This PR contains candidate capture and replay machinery for transcribing CDC streams for differential dataflow collections.

The documentation speaks about the protocol, which is designed primarily to tolerate duplication and reordering, and to allow as much parallelism as is easy to accomplish.

The code has been lightly tested, and then changed, and then not tested. It could certainly use more testing before anyone relies on it!