Shrink upsert allocation (#280)
This PR removes a memory leak in the upsert priority queue, which would never reduce its size even when drained. If there is sufficient waste (a factor of four) the queue will now scale down its allocation to fit.

There are more performant ways of doing this with a custom heap implementation (as a sequence of geometrically sized allocations) but for the moment let's just start by returning the memory.