agent: Avoid temporary allocations by keeping a temporary buffer
This commit changes the TraceAgent to keep a buffer for temporary
antichains around, which is used for setting the logical and physical
compaction. The upside is that it avoids temporary allocations, but
on the downside it keeps another antichain around.

This should be fine, unless there is a defective use case where the size
of the antichain grows significantly and the potentially unused memory will
not be reclaimed anymore.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>