reduce_core: present referenced data to logic function
`reduce_core` was previously forced to clone the output data before
presenting it to the logic function for inspection.

Simply annotating the type with a reference was not possible due to the
buffer being held in the same struct as some of the data it contained,
making it a self-referencial struct which the borrow checker doesn't
allow.

However, the only reason this buffer was held in the same struct was to
reuse its allocation. This patch therefore creates a brand new vector
every time we need to process data that contains references instead of
cloned data. The expectation is that in the general case it's better to
allocate a vector of references than clone a bunch of values.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>