Extract ready times from existing chains
Teaches the merge batcher to extract ready times from the existing
chains, and maintaining the chain invariant after extracting data. This
reduces the effort to maintain data that is not yet ready, by
maintaining a frontier per chain block that allows us to efficiently
decide that a block needs to be inspected or not.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>