Split default merge batcher
This change splits the default merge batcher implementation into a type
that maintains the outer part of its algorithm, specifically knows how
to maintain chains, and an inner part that knows how to maintain the
individual batches in chains. The benefit is that the outer part does
not need to know about the contents of the containers it holds on to
because that's encapsulated in the inner trait's implementation.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>