Remove reborrow
Removes the reborrow functionality to locally shorten lifetimes. It's a bit
unclear to me whether this is a change worth making or not, but it
certainly simplifies a few things, at the expense of requiring different
PartialEq implementations, namely for references of two different
lifetimes, which does not match the regular implementation. Tests look
green, but we might not have perfect coverage.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>