Separate chunker from batcher
The chunker was part of the batcher and responsible for transforming input
data into the batcher's chain format. Hence, the batcher needed to be aware
of its input types, although it would not otherwise use this information.

With this change, the chunker is separate of the batcher. This simplifies
the logic within the chunker slightly, but most importantly moves the
responsibility to form chunks to whoever holds the batcher. In
Differential, this is `arrange_core`. It now learns about an input
container type and a chunker and uses the chunker to convert input data
to chunks of sorted and consolidated data.

Signed-off-by: Moritz Hoffmann <antiguru@gmail.com>